% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.R
\name{AD}
\alias{AD}
\title{AD: Average Deviation Within-Groups}
\usage{
AD(x, grpid, model, scale)
}
\arguments{
\item{x}{Item on which to estimate AD}

\item{grpid}{Grouping/clustering variable}

\item{model}{User-supplied description of multilevel measurement model (e.g., consensus)}

\item{scale}{Minimum and Maximum Values of the Scale}
}
\value{
Estimates AD
}
\description{
This function estimates Burke, Finkelstein, & Dusig's (1999) within-group agreement statistic
for a single item, denoted AD.
}
\details{
The input consists of a data frame, and the names of the grouping variable and item.
The output consists of a data frame containing the group names, the group means, group medians, and the
estimates of AD(j) about means and medians.

Link to Burke, Finkelstein, & Dusig (1999):
}
\examples{
data(lq2002, package = "multilevel")
AD(x = lq2002[ ,c(3:13)], grpid = lq2002$COMPID, scale = c(1,5), model = "consensus")
}
