% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGGREGATE.DATA.R
\name{AGGREGATE.DATA}
\alias{AGGREGATE.DATA}
\title{AGGREGATE.DATA: aggregate data and merge into combined data frame}
\usage{
AGGREGATE.DATA(grpid.name, x.name, data, aggr.stat)
}
\arguments{
\item{grpid.name}{Name of the grouping/clustering variable}

\item{x.name}{Name(s) of the item(s)/variable(s) to be aggregated}

\item{data}{Name of data frame containing group and vars}

\item{aggr.stat}{Name of the function used to aggregate data}
}
\value{

}
\description{
This function augments the aggregate.R base function.
}
\details{
The input consists of a data frame, and the names of the grouping variable and item.
The output consists of a data frame containing the group names, the variance within each group, and
estimates of rwg based on multiple null response distributions (see LeBreton and Senter, 2008).
}
\examples{
data(lq2002, package = "multilevel")
AGGREGATE.DATA(grpid.name = "COMPID", x.name = c("LEAD", "TSIG", "HOSTILE"),
               data = lq2002, aggr.stat = "mean")

df.aggr
df.combined
}
