% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWG.R
\name{RWG}
\alias{RWG}
\title{RWG: Within-Group Agreement}
\usage{
RWG(x, grpid, model, scale, reset = F, cutoff)
}
\arguments{
\item{x}{Item on which to estimate RWG}

\item{grpid}{Grouping/clustering variable}

\item{model}{User-supplied description of multilevel measurement model (e.g., consensus)}

\item{scale}{Minimum and Maximum Values of the Scale}

\item{reset}{Logical option for handling negative estimates of RWG; FALSE retains negative values; TRUE resets values to 0}

\item{cutoff}{User-supplied cutoff for justifying data aggregation}
}
\value{
List containing results of analysis
}
\description{
This function estimates James, Demaree, & Wolf's (1984) within-group agreement statistic
for a single item, denoted rwg.
}
\details{
The input consists of a data frame, and the names of the grouping variable and item.
The output consists of a data frame containing the group names, the variance within each group, and
estimates of rwg based on multiple null response distributions (see LeBreton and Senter, 2008).

Link to James, Demaree, & Wolf (1984):
Link to LeBreton & Senter (2008):
Link to LeBreton, Moeller, & Senter (2021):
}
\examples{
data(lq2002, package = "multilevel")
RWG(x = lq2002[,c(3)], grpid = lq2002$COMPID, model = "consensus",
    scale = c(1,5), reset=FALSE, cutoff=.50)
}
