% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGA.R
\name{WGA}
\alias{WGA}
\title{WGA: Within-Group Agreement}
\usage{
WGA(x, grpid, scale, model, reset = F, cutoff = 0.5)
}
\arguments{
\item{x}{A vector representing a single item or a matrix representing a scale of interest.
If a matrix, each column of the matrix represents a scale item, and each row
represents an individual respondent.}

\item{grpid}{A vector identifying the groups from which x originated.}

\item{scale}{A vector with the lower and upper response options (e.g., c(1,5))
for a five-point scale from strongly disagree to strongly agree.}

\item{model}{User-supplied description of multilevel measurement model (e.g., consensus)}

\item{reset}{Logical option for handling negative estimates of RWG; FALSE retains negative values; TRUE resets values to 0}

\item{cutoff}{User-supplied cutoff value for justifying data aggregation}
}
\value{
Estimates of within-group agreement
}
\description{
This function computes multiple estimates of withing-group agreement including:
 James, Demaree, & Wolf's (1984) RWG(J), Burke, Finkelstein, & Dusig's (1999) AD(J),
 and Brown & Hauenstein's (2005) AWG(J).
 Estimates of RWG(J) are computed for multiple null distributions.
}
\details{
The input consists of a data frame, the names of the grouping variable, and the
names of items.  The output consists of a data frame containing the group names,
group sample sizes, the variance within each group, and
the various estimates of agreement.
}
\examples{
data(lq2002, package = "multilevel")
WGA(x=lq2002[,3], grpid = lq2002$COMPID, scale=c(1,5), model = "consensus", reset = FALSE)
}
