% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWGJ.R
\name{RWGJ}
\alias{RWGJ}
\title{RWGJ: Within-Group Agreement (multiple items)}
\usage{
RWGJ(x, grpid, model, scale, reset = F)
}
\arguments{
\item{x}{Items to use when estimating RWGJ}

\item{grpid}{Grouping/clustering variable}

\item{model}{User-supplied description of multilevel measurement model (e.g., consensus)}

\item{scale}{Number of scale points ranging from 5 to 11}

\item{reset}{Logical option for handling negative estimates of RWG;
FALSE retains negative values; TRUE resets values to 0}
}
\value{
List containing a matrix of the null error values, estimates of RWG, and
        summary of how many negative values were obtained
}
\description{
This function estimates James, Demaree, & Wolf's (1984) within-group agreement statistic
for a multi-item scale, denoted RWGJ.
}
\details{
The input consists of a data frame, and the names of the grouping variable and item.
The output consists of a data frame containing the group names, the variance within each group, and
estimates of rwg based on multiple null response distributions (see LeBreton and Senter, 2008).

Link to James, Demaree, & Wolf (1984):
Link to LeBreton & Senter (2008):
Link to LeBreton, Moeller, & Senter (2021):
}
\examples{
data(lq2002, package = "multilevel")
RWGJ(x = lq2002[,c(3:13)], grpid = lq2002$COMPID, scale = 5, reset = F)

}
